requirements.txt
# requirements.txt
weasyprint>=53.0
jinja2>=3.1.2
beautifulsoup4>=4.10.0
requests>=2.28.1
python-telegram-bot>=20.0
pillow>=9.2.0
python-dotenv>=0.20.0
deep-translator>=1.10.1
aiohttp>=3.8.3
pathlib>=1.0.1 